Hilt DI added + minor polish

+++ app/src/main/java/com/pashacabu/tmdb_app/model/data_classes/room_db_tables/MovieDetailsEntity.kt
@@ -19,7 +19,7 @@ data class DBMovieDetails(
     @ColumnInfo(name = "backdrop")
     var backdropPath: String? = null,
     @ColumnInfo(name = "revenue")
-    var revenue: Int? = null,
+    var revenue: Long? = null,
     @ColumnInfo(name = "genreName")
     var genreName: String? = null,
     @ColumnInfo(name = "genreID")
